Replace string type check in get_subclasses with try/except
The string check in `helpers.get_subclasses` was preventing unicode
strings from being used. Furthermore, type checking on instances isn't
very pythonic, so the replacement is a try/except around the use of the
argument. It's worth replacing the check, rather than simply removing
it, to offer a more helpful error message.

Python 2 and Python 3 raise different exceptions in this case, so the
except clause accounts for both.