Fix launch requirements
Previously, when launching an environment, a stack would be launched if
all parent stacks were not in the *_FAILED state. This commit changes
this to only launch a stack if all parent stacks are in the *_COMPLETE
state. These stricter requirements fix issues which occur when a parent
stack is in an e.g. DELETE_FAILED state.