Add simplified Config and ConnectionManager inits
Config and ConnectionManager both require quite a lot of
context to initialise. This commit adds new functions for
initialising them. The new function require only the sceptre
directory, an environment path or stack name. This should
make it easier to create config and connection managers
without having to instantiate an environment or stack
first. All new functions implement memoization. If a
particular Config or ConnectionManager is requested twice,
a cached version is returned the second time. The functions
aren't being used at the moment, but we should note that they
aren't necessarily thread safe.