Resolve #198
Auto create or update stacks with changesets

Change sets are typically used for reviewing changes to an existing
stack before execution. However, they also support creating new stacks
via the --change-set-type=CREATE argument (which defaults to
--change-set-type=UPDATE). Sceptre does not support this parameter and
relies on the default.

There may be other compelling use cases for this, but a necessary one
is to be able to deploy SAM (Serverless Application Model) templates via
Sceptre. SAM templates are a great way to configure Lambda+API Gateway
services. For some reason, create-stack cannot be used with templates
that have a Transform directive; the workaround is to create a change
set and execute it with --change-set-type=CREATE to create the stack,
which works with SAM-transformed templates.

This commit automatically determines which change set type to use when
creating a change set.

The stack_exists method is a fork from awscli's has_stack method, from
https://github.com/aws/aws-cli/blob/bb92a4040900939e8d535923fe8bf3684c574a85/awscli/customizations/cloudformation/deployer.py#L36-L70