Remove 'Response Metadata' from CLI output
Resolves #84

We need to improve the CLI output. The Response Metadata information
returned from Boto clutters the success message when the
validate-template command is called from the CLI.

The validate template tests for the CLI  are also updated to reflect
the expected changes in output. A new test is added to test invalid
templates.