Fix path creation when encountering stacks.
When constructing the path of encountered nodes, they were inserted
at the beginning of the list. However, when that list is trimmed
down upon finding a cycle, the logic suggested that the nodes
should have instead be added to the end.

This commits now adds nodes to the end when constructing a path.