Add Estimate Template Cost Functionality
Resolves #106

We need to be able to estimate the cost of our Cloudformation templates
before creating them. AWS provides a convenient API to make this happen.
This API has been implemented in Stack and the Cli to enable the use
within python modules and the CLI.

If called from the CLI, a unique URL containing the estimated cost
will be printed to stout and will attempt to open a web browser with
that URL.