[Resolve #310] Pass `profile` key to ConnectionManager
According to the docs
(https://sceptre.cloudreach.com/latest/docs/environment_config.html#profile),
it should be possible to specify a `profile` name in a config.yaml to
indicate which AWS CLI profile to use (in ~/.aws/credentials) to fetch
credentials.

However, the `profile` parameter wasn't getting passed to the Stack
constructor, which then needs to pass it along to the ConnectionManager.