[Resolves #382] Remove forked PR integration tests filter
It is not entirely clear why a filter was added to prevent integration
tests running for forked PRs. Recently, we've had to revert a
number of merges after discovering they don't pass integration tests.

This filter is being removed with the intention that it will provide
visability of integration tests for forked PRs before they are merged
into master.