[Resolves #257] improved Cross Region/Account Support (#395)
This will deprecate the original functionality of using iam_role and instead allow the user to use an AWS Profile (supported by boto and awscli). The profile itself is usually defined in the ~/.aws/config file (prefixed with profile <name>) but could be placed anywhere with the use of AWS_CONFIG_FILE which specifies the path to the configuration file.

This will break any sort of compatibility with the use of iam_role and instead a profile has to be created. The profile can be added to the sceptre configuration file in the same manner as the iam_role used to.

The current level of precedence is as follows (and is all according to the default AWS Config/CLI behaviour linked below):

Environment variable credentials take priority over profiles
Region specified in the Sceptre Stack configuration will take priority over the one specified in the profile
Profile takes priority last
You are able to specify a profile which can target a different account and region while providing credentials through environment variables, instance role or else as required. More information on AWS profiles, configuration and credentials can be found here.

Example stack configuration:

profile: development-profile
project_code: cost-code
As far as the the YAML / sceptre resolver tag are concerned:

!stack_output should work as intended with full cross account/region functionality.
!stack_output_external has been redesigned slightly, it can now take in an optional 2nd arg, which is the profile. Example: SomeParam: !stack_output_external SomeStack::Output SomeProfile where SomeProfile is the profile defined in your AWS Config file to use for fetching that specific stack output. If the last argument isn't supplied then it is assumed to use whatever default is available..