[Resolves #298] Fix stack config dependencies merging
Previously, the dependencies data structure in stacks were not being merged properly. 
This resulted in issues, such as #298, being reported. For example, in the sceptre project 
displayed below, if bundle/b/y.yaml had a !stack_ouput from bundle/a/x.yaml an implicit 
dependency is created. If we then specified that bundle/b/y.yaml had a dependency on 
bundle/a/x.yaml the dependencies would not resolve properly.

```
.
├── config
│   ├── bundle
│      ├── a
│      │   └── x.yaml
│      ├── b
│      │   └── y.yaml
│      ├── c
│      │   └── z.yaml
│      └── config.yaml
│   
├── templates
    └── template.json

```

The fix submitted with this commit gives us flexibility in dealing with different merge strategies for config items if required in the future.