[Resolves #160] Improve delete-change-set message (#449)
There does not appear to be a suitable exception returned by AWS when
attempting to delete a change-set that does not exist. The AWS API
returns a 200 regardless of whether the change-set existed or not.

Therefore the best we can do is signal to the user that if the
change-set existed it has been deleted.