[Resolves #187] Fix launch env dependency path
As highlighted in #187, a user must specify a full path for stack
dependencies, even within the same environment. The documentation
shows the initial intention was to allow a user to reference the
stack name without the full path.

Given the following project structure:

config/
- config.yaml
- dev/
  - vpc.yaml
  - subnet.yaml
- templates/
  - vpc.yaml
  - subnet.yaml

Previously, a user had to specify dependencies in one of two ways:

``
dependencies:
    - dev/vpc
    - {{ environment_path.0 }}/subnet
``

This commit means that users can now specify dependencies as:

``
dependencies:
    - vpc
    - subnet
```