Add SceptrePlan and Executor
This commit is part of a wider piece of work that changes the way
Sceptre will load and execute commands on stacks and stack groups.
Between this commit and bf39d57f1591911799aad985f8e748fb7e2b8b4b sceptre
will be able to preload all the objects and configuration neccessary to
execute an action, rather than taking actions on-the-fly, which is the
behaviour in version 1.