Add SceptreContext and use SceptreContext in CLI
Currently context is passed in through the CLI and then handled in an
adhoc fashion throughout the application.

As part of refactoring the way config is handled in Sceptre and in an
attempt to make the API more consistent the concept of SceptreContext
has been added.

A SceptreContext provides information to a SceptrePlan about the
environment in which the SceptrePlan is being executed.

A SceptreContext contains information such as the project path, the
current command path, config path, user variables, options, and
formatting settings.

A SceptrePlan is responsible for resolving all Stack and StackGroup
configurations, dependencies and launch order before being executed by a
SceptrePlanExecutor. Access to the various paths is now much cleaner.