Fix stack selection when deleting stack groups
When deleting a stack group we want to delete any stacks in the group,
but also any stacks that depend on those stacks. In order to identify
these stacks properly the configuration for all stacks is now loaded
into the StackGraph, and the launch order is computed by finding all the
stacks reachable from the stacks targetted by the given command path.

For deletion, this search is performed on the reversed graph, meaning
anything that depends on the stacks to be deleted will also be deleted,
as desired.