[Resolves #522] Add ignore-dependencies flag
Adding a `--ignore-dependencies` flag so that Sceptre will only execute
a command on the given Stack or StackGroup and ignore all other
dependant Stacks. Useful when wanting to operate on a single Stack, for
example, when creating a Change Set.

Reverts `ConfigReader().construct_stack()` to a previous implementation
(182f0b64f1bfbb439c7e6f7778a3ce0d0774d187) to handle case where root is
the stack file and changes root to the command path so that full project
is not loaded upfront.