[Resolve #579] fix update-stack return code
The update-stack command returns a non 0 status when there are no
changes to peform.  No updates is not be an Error, it should behave
like launch-stack where no updates is not an error either.  Fix
it so that sceptre ignores the error from boto when there are no
changes to the stack.  This will help run sceptre in a CI/CD scenario.