Fix stack_output recursion error
There seemed to be a recursion issue which was probably an issue with
when a ResolvableProperty was called.

This commit attempts to fix this, however there may still be an underlying
issue. If there is then the real fix will be to try and remove
ResolvableProperties in favour of another implementation, they seem to
be causing more issues than they are worth.

[Resolves #558]