Fix `--ignore-dependencies` with `stack_output`
This commit fixes an issue where a users attempts to launch a single
stack that utilises the `stack_output_resolver` with the
`--ignore-dependencies`.

Previously, with `--ignore-dependencies` set, any dependencies that arose
through `!stack_output_resolver` would not be added to
`stack.dependencies` and the `next()` function in
`StackOutput.resolve()` would raise a `StopIteration` as the
stack.dependencies list would be empty.

This commit handles the stack_output by utilising the argument passed
into `stack_output`.

[Resolves #778]