[Resolves #571] Add iam_role capability for stack-config
This allow users to specify `iam_role` in their stack config.
The IAM role is evaluated at the very end after the environment
variables and profile for AWS.

The following order of precedence applies for the stack config:
- Environment variables for AWS are used (access/secret/session)
- The `region` set in the stack config is used
- The `profile` set in the stack config if used
- The `iam_role` is assumed using the current session available

If a user has environment variables, region, profile and iam role
set then it will use the environment variables to create a session
using the defined region in the stack config. It will then assume
the role using that session. The credentials from the profile are
ignored in this case. If environment variables are not present
then the profile will be used if available, otherwise the default
local credentials will be used.

In addition, the stack output external resolver has been modified
to allow for additional arguments separated by double colons (::).

The format now supports profile::region::iam_role and is fully
compatible with current and older versions

The following are all valid examples:
- !stack_output_external Stack::Output profile::eu-west-1::role_arn
  - Creates session from profile, assumes role and uses region
- !stack_output_external Stack::Output ::eu-west-1::role_arn
  - Assumes the role from the default credentials and uses region
- !stack_output_external Stack::Output profile
  - Uses profile, profile can include role and region as well
- !stack_output_external Stack::Output ::::role_arn
  - Assumes the role from the default credentials