Detail how to use Sceptre hooks to deploy lambda
In order to deploy Lambda functions using either the AWS CLI
or SAM CLI we need to generate the packaged version of the
Lambda functions first. This patch details how to add the "package"
command using Sceptre hooks in a way that reuses existing Sceptre
config.yaml file information to populate the command details.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>