Disable integration tests
Disable integration tests because we don't have access to the
AWS account it depends on.  We'll need to get an account before
we can run these tests again.

The intention is to disable it to get the builds passing again and
to be able to merge non-code changes. We also inted to get these
tests running again before merging any code related changes.