Add info to configure Sceptre's API Access
The docs seem to imply that the user knows that Sceptre uses the same
configuration as the AWS CLI and thus does not provide details on how
to set it up for the first time.

This patch clarifies that by explicitly noting to use the AWS CLI to
configure the configuration files and links to the source docs.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>