Set up a basic pyproject.toml for PEP-518 builds (#1009)
This made a number of changes, working toward a singular goal: Implementing a `pyproject.toml` to allow for PEP-518 builds. This will allow for better testing in tox, the possibility of using other packaging systems, and this file is also the only one where config for `black` can be stored. In addition to that, a lot of projects are now allowing you to put their config into this file. Such as I have done with pytest!

A summary of my commits:

- Ignore .venv folders
- Don't force Python development versions
  - I removed the `.python-version` file as it was forcing developers to use `3.6-dev` and `3.7-dev` versions of Python which are very old and probably not good to be using at this point.
- Read the version from the file instead of importing it
  - With the way that PEP-518 builds work, the project is put into an isolated environment before being built. In that isolated environment, nothing is installed, so the `from sceptre import __version__` call was failing. Now, that file is scraped manually for the version.
- Upgrade to the latest pytest
  - This was 3 major versions behind, and this latest allows for placing config in `pyproject.toml`!
- Add a basic pyproject.toml & shift pytest config over