Add new badges & clean up README structure (#1031)
The main goal here was just to add some more informational badges to the README. Since we're not currently using any kind of "code quality" service, none of those are available right now. But those can easily be added in the future if that situation changes.

While I was in the README, I fixed the few `markdownlint` warnings that popped up. I debated on adding `markdownlint` to the `pre-commit` toolchain, but I hesitated as there is not an active project written in Python and I did not want to add another language to the build toolchain. If that is wanted however, we have the following options:

* [Run the canonical Ruby linter](https://github.com/markdownlint/markdownlint/blob/master/.pre-commit-hooks.yaml#L1-L6)
* [Run the canonical Ruby linter in Docker](https://github.com/markdownlint/markdownlint/blob/master/.pre-commit-hooks.yaml#L8-L13)
* [Run a Javascript linter](https://github.com/igorshubovych/markdownlint-cli/blob/master/.pre-commit-hooks.yaml)