[Resolves #1090] Install troposphere as an extra package (#1104)
* [Resolves #1090] Install troposphere as an extra package

Scepter supports the deployment of troposphere templates however the
troposphere library does not get installed on a sceptre install because
the library dependency is not setup. This means that troposphere must
be installed separately from sceptre in order to use it. The problem
with that is that users may install a version of troposphere that may
not work with sceptre. To fix this issue I think we should add a versioned
troposphere dependency to sceptre so that users will get a version that
is tested and will most likey work with the installed version of scepter.

* fix linter error

* fix typo