[Resolves #213] Add support for template handlers  (#1088)
This PR is an update to PR #864

This PR adds support for template handlers, a new entity used to load templates from external sources. The `Template` class now always expects a configuration object for the template handler, and will instantiate it based on the entry points found for `sceptre.template_handlers`. In this way, others can add new template handlers and use them in the Sceptre config.

`TemplateHandler` objects define a JSON schema that will be validated upon use. This will make sure that template handler authors and users provide the correct configuration.

An example of the S3 template handler is as follows:

```yaml
template:
    type: s3
    path: my-bucket/my-fancy-template.yaml
parameters:
    ...
sceptre_user_data:
    ...
```

All the properties of the complex `template` object will be passed into the template handler as arguments (except for `type`). 

To not break backwards compatibility, the `template_path` property can still be used. It will be wired into a `file` template handler in `stack.py`, so the 'ugliness' is contained there. `Template` can then always work with handlers, without considering `template_path`. The `file` template handler contains all the logic for loading Jinja2 templates and/or Python scripts, that was moved from `Template`. So that logic is not in the core of Sceptre any more (which is good IMHO).

The `s3` template handler that is included here does not support Jinja2 templates or Python scripts. That is a limitation that can be solved by writing some code that will check the extension and parse the S3 contents from memory (similar to what `file` does, but that only supports files on disk, not strings in memory).

Co-authored-by: Léon Rodenburg <l.rodenburg@instanceof.nl>
Co-authored-by: Lucas Vieira <lucas.vieira94@outlook.com>