Update troposphere library
When attempting to deploy troposphere template I get the following error

```
AttributeError: 'Template' object has no attribute 'to_yaml'

```

Accoring to https://github.com/cloudtools/troposphere/issues/759  it
seems like we need to update trophoshere to a version where the
`to_yaml` was added back.  The current version of trophosphere is
pretty old anyways.

The latest version is 3.0.3 however updating to 3.0.3 causes unit
tests to fail so we are sticking to the latest 2.x version.