[Resolve #1138] Bugfix for j2_environments (#1137)
Before this, the j2_environment feature that was added in #966 had been broken during the recent implementation of the `TemplateHandler` feature. The root cause was that the `_render_jinja_template` had been rewritten as `render_jinja_template` in `sceptre/template_handlers/helper.py` but the logic for J2 environments had not been copied.

The main change here is to hook the `stack_group_config` attribute of the `Template` class into this helper. 

As part of this I also cleaned up the old `_render_jinja_template` function which was no longer used and its tests.