[Resolve #1135] Fix path to templates (#1141)
Currently the root file path for the template file handler points to a different
location than the existing template_path. This makes the migration process
confusing and arduous for users. Therefore when a user specifies a relative
path to the file we make both parameters point to the same location, the
`sceptre_project_dir/templates` directory.