[Resolves #1148] Correct path logic
This resolves a bug introduced in 5c0ab39. Note also that prior to
introduction of template handlers, a hack had existed to solve the
problem of computing the path when `--dir` is specified, see
https://github.com/Sceptre/sceptre/blob/354e87e8fbf2dafd2692c7c87abc1ecc35ea3f40/sceptre/template.py#L152-L167

In the case of --dir, the `File` handler class's `handle` method has
received a path relative to the `project_path` due to the previously
incorrectly-named method in the `ConfigReader` class
`_get_absolute_template_path` returning the relative path.

This patch renames the incorrectly-named method
`_get_absolute_template_path` as `_get_relative_template_path` and
corrects the logic in the `File` class's `handle` method.