[Resolves #1201] Fix dependency conflict (#1206)
A change to MarkUpSafe caused Sceptre to stop working.  This is due to
Sceptre's dependency on Jinja2 ver 2.x.  The recommeneded solution to
resolve the conflict is to update Jinja2 to ver 3.x.

Updating Jinja2 also required updating moto and some tests

Python 3.6 is EOL[1] so we are dropping support for it.

[1] https://endoflife.date/python