[Resolve #1307] Fixing Connection Manager bug (#1308)
This resolves a bug discovered in the ConnectionManager from #1307.

The specific situation is when the ConnectionManager.call is invoked using a stack_name and explicitly passes None for each of profile, region, and sceptre_role. The overhaul of the ConnectionManager in v4 accidentally omitted this edge case which happen to specifically affect the !stack_output_external resolver when it's run without either an AWS_DEFAULT_REGION environment variable or without an AWS_PROFILE environment variable set to a profile that defines a region.