[Resolve #1336] Add bugfix in Stack Differ
In the case of a parameter dictionary that contains nulls (None) the
Boto3 library has been found to take care of converting None to "". This
means that launching such a stack in CloudFormation with a parameter
dict containing params set to None works fine, whereas this differ
raised an exception when it tried to strip newlines.

This patch stops newlines being stripped in the case of a parameter that
is set to None.