[Resolve #1291] Fix attribute error caused by a relative project path (#1296)
Resolve [#1291](https://github.com/Sceptre/sceptre/issues/1291)

Given a relative project path with a leading `.` and duplicate dependencies in multiple stack configurations, there is a chance of an AttributeError being raised. This is caused by dependencies failing to resolve into a valid Stack object due to an invalid path.

This PR fixes the issue by ensuring a given project path is resolved to the absolute path when initialising the Context.

An additional test of context with a relative path is included to ensure correct resolution.