[Resolve #1358] Updating PyYaml to ^6.0 (#1359)
There's an issue where the latest version of Cython that just dropped today (v3) is incompatible with PyYaml v5 on certain versions of Python (e.g. 3.10). 

This was raised in #1358. We probably should do some preliminary testing with Sceptre on this version of PyYaml, but it theoretically fixes the install issue. 

According to comments on [AWS CLI's issue about this](https://github.com/aws/aws-cli/issues/8036#issuecomment-1638544754), the issue has to do with the build process when there is no pre-compiled wheel for your platform/OS. PyYaml v6 has precompiled wheels for Py3.10 and Py3.11. **Unfortunately, if there is no precompiled wheel for your platform, you will still encounter this issue.**