[Resolves #1213] Add shell option to cmd hook (#1372)
The cmd hook now accepts named arguments `run` and `shell` to run a command in a
given shell. For backwards compatibility it still supports the unnamed
argument to run a command in the default shell.

The documentation shows the expected output for examples of each argument style.

The unit tests no longer depend on the implementation of Cmd and test only the
behavior. This makes it easier to change how the hook controls the subprocess.