[Resolves #1383] Simplify stderr test for cmd hook
The old test expected `/bin/sh` to write Dash's error message. This works on
Ubuntu 20.

It old test failed  when `/bin/sh` pointed to `/bin/bash`. This happens on
macOS.

The new test no longer depends on system-defined side effects of a missing
command. It just writes to the standard error stream and checks that the same
text comes out.