Update documentation for docs (#1403)
Update the docs to show contributors how to generate Sceptre docs using poetry instead of make. The CI does not run the `make` command and neither do developers anymore.  All build, test, and deployment tasks use poetry and all the tasks have been documented therefore in the interest of making the project less of a burden to maintain we remove the makefile.

Note that the docs/Makefile is used to generated docs and that one is still needed, the makefile in the root folder is the one that's no longer needed and is being removed. 

depends on #1398