Generate github test matrix from tox
We only want one static list of environments to run unit tests
and that list should come from tox.ini

This sets up tox-gh-matrix[1] to dynamically generate the list
of test environments for github actions to run. It will allow
us to consolidate the list to just one for easier maintenance
going forward.

[1] https://github.com/medmunds/tox-gh-matrix