[Resolve #1468] Cast parameters to strings
In already-resolved and merged issue #1443 we implemented parameter
validation.

Subsequent discussion led to a consensus that it would be better to not
only validate parameters, but also cast where applicable to string. This
will lead to a cleaner user experience for most Sceptre users where
parameters passed as ints, floats and bools in their YAML values files
are cast to strings first and then sent to CloudFormation, since
CloudFormation's API is unable to handle other types.

This PR changes the behaviour of the `_ensure_parameters` method in the
Stack class to handle casting to strings where applicable instead of
only validating.