[Resolve #915] Support ChangeSetType CREATE (#1469)
This adds logic supporting the case of creating change sets for a stack that does not already exist. That is, it adds support for change sets of ChangeSetType CREATE. https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/cloudformation/client/create_change_set.html

To add some context, there is no obvious reason why Sceptre does not already support change sets of type CREATE since the logic for detecting stack status already exists for the launch command. The project's initial commit already has this inconsistency and essentially the same code. It is probably an oversight.

The changes have already been tested by Rohit S in this Slack thread
https://og-aws.slack.com/archives/C01JNN8RGBB/p1716869822031499

In addition, it is necessary to add logic to handle the REVIEW_IN_PROGRESS state, both for the case of the launch and create action.
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-describing-stacks.html

Creation of a change set on a non-existent stack actually creates a stack and places it into the REVIEW_IN_PROGRESS state. If Sceptre is then to launch this stack, it is necessary to firstly delete the stack, just as it already does in the cases of states CREATE_FAILED and ROLLBACK_COMPLETE.