[Resolve #1483] Handle errors in change sets
This adds code to handle the case of failures in stack groups that have
missing or errored stacks when creating, describing and executing
change-sets.

Before this, any stack group that contained a stack that had been
deleted, or had errored out, resulted in all change set operations on
that stack group failing.

The reason for this is no error or exception handling had ever been
added for change sets, but assumed all stacks in the stack group were
clean and could respond to the change set APIs.

For example, deleting change sets might error out with:

```
% sceptre delete -y network.yaml sceptre-network
The Change Set will be delete on the following stacks, if applicable:
network

"An error occurred (ValidationError) when calling the DeleteChangeSet
operation: Stack [test-e2e-direct-connect] does not exist"
```

This is particularly an issue for sites that employ as best practice
analysis of change sets for safety prior to launching updates.

This patch adds the missing error handling for `delete_change_set`,
`create_change_set`, `describe_change_set` and
`execute_change_set`.