[Resolve #1506] Remove unused __eq__ method
Way back in 2018, the original team were adding the debug command, and in
that context set about to improve the output in some debug circumstances
during inspection of Sceptre's internal objects.

In response to https://github.com/Sceptre/sceptre/issues/570, a custom
`__eq__` method was added in df9fc201aaed59c173b11b80e804244a5b2f3bfb.

Based on what can be seen of the original implementation of the custom
`__eq__`, it has not been maintained in a long time, other than changes
to it that appeared to cause issues for the current maintainers.

--

A recent change to improve the output during cyclical dependency errors
has introduced a call to `nx.find_cycle`. This function explicitly
searches for a cycle in the graph. This appears to resolve in code
testing for equality of nodes and thus calls to our custom `__eq__`
function.

The `__eq__` method however has been technically broken ever since
`template_path` was made optional. This then results in `__eq__`
failing. This fails as the code to indicate that `template_path` is
deprecrated is then executed, which will fail if the setting has not
really been defined.

--

This PR proposes to simply remove `__eq__` since it is believed that it
is no longer in use by anything. It has been more than 2 years since
`template_path` was deprecrated and made optional, and the fact of this
bug having never been observed until now is good evidence that it is not
used or needed.