[Resolve #1482] Make shell optional in cmd hook (#1554)
Now you can set the command to run as an object without also having to set the shell.

In this case it uses the default shell, as if the command to run were passed as a string.

This allows you to use resolvers to set the command without worrying about which shell to invoke. See @dmarra's use case in #1482.